nlp_worker/requirements.txt
# Core
requests
python-dotenv

# Async (if you decide to fetch multiple sources in parallel later)
aiohttp

# Optional: if using Reddit via praw instead of RapidAPI
praw
